requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
cors
python-dotenv
bcrypt
groq
